+++ lib/google_maps.dart
@@ -1,0 +1,30 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+
+class GoogleMapWidget extends StatefulWidget {
+  @override
+  _GoogleMapWidgetState createState() => _GoogleMapWidgetState();
+}
+
+class _GoogleMapWidgetState extends State<GoogleMapWidget> {
+  Completer<GoogleMapController> _controller = Completer();
+
+  static const LatLng _center =
+      const LatLng(62.238838951526844, 25.75458690343214);
+
+  void _onMapCreated(GoogleMapController controller) {
+    _controller.complete(controller);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GoogleMap(
+      onMapCreated: _onMapCreated,
+      initialCameraPosition: CameraPosition(
+        target: _center,
+        zoom: 11.0,
+      ),
+    );
+  }
+}
