+++ lib/main.dart
@@ -1,6 +1,6 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'bottom_sheet.dart';
+import 'google_maps.dart';
 
 void main() => runApp(MyApp());
 
@@ -10,15 +10,6 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  Completer<GoogleMapController> _controller = Completer();
-
-  static const LatLng _center =
-      const LatLng(62.238838951526844, 25.75458690343214);
-
-  void _onMapCreated(GoogleMapController controller) {
-    _controller.complete(controller);
-  }
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -36,42 +27,8 @@ class _MyAppState extends State<MyApp> {
         ),
         body: Stack(
           children: <Widget>[
-            GoogleMap(
-              onMapCreated: _onMapCreated,
-              initialCameraPosition: CameraPosition(
-                target: _center,
-                zoom: 11.0,
-              ),
-            ),
-            DraggableScrollableSheet(
-              initialChildSize: 0.05,
-              minChildSize: 0.05,
-              maxChildSize: 0.9,
-              builder:
-                  (BuildContext context, ScrollController scrollController) {
-                return Container(
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.vertical(
-                      top: Radius.circular(16),
-                    ),
-                  ),
-                  child: SingleChildScrollView(
-                    controller: scrollController,
-                    child: Column(
-                      children: [
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Icon(Icons.keyboard_arrow_up),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              },
-            ),
+            GoogleMapWidget(),
+            BottomSheetWidget(scrollController: ScrollController()),
           ],
         ),
       ),
